requirements.txt
pandas
tweet-preprocessor
spacy
gensim